Rework SVG traffic sign files as part of the package
Internal package Wiki:
- Remove the download of SVGs, that was just to test out this part of the process
- Improve and fix the app's wiki page

New internal package SVGs:
This one downloads the SVGs based on our traffic sign data.
It is already prepared to work for multiple countries.
We have to copy the files from here to the package manually (as part of the build).
I did not manage to put this into the package because the download and "write file" code interfered with the react app; I didi not manage to exclude parts of the package which is why it is separate now.

Package:
The sign data does not know the name of the sign, but there is a helper that will create the right name and we can use this to import.

App/NextJS:
Importing the SVGs from the packages was tricky, but this seems to work…

Multi-Country:
Quite a few changes are related to make the app ready to add more countries.

SVGs:
All SVGs are now optimized and have som accessibility helper inside.