Tool: Refactor to fix NextJS 15 build erros; use context instead of params for language
The NextJS 15 build failed with hard to understand errors that I assume have to do with the fact that this app is build statically but NextJS is not really made for that. Those use cases are just not that well tested and defined.

We use page params to share the country code between components. This is refactored to use a context setup.

Also the pages are extracted into components. Other languages need to create a folder like /NL/* and add pages like what we have for /DE/*. The advantage is, that some pages are not relevant for other countries because they are DE specific so this setup brings us a big closer to cleaning this up.

This also enables the react compiler. But it is not clear if that actually works or what the effects are.