Signs: Fix sign svgs being broken with custom values
Once we changed the custom value, the SVG would be broken because it was looking for a sign with this custom value in the file name.

We need to have those values part of the file name to be able to duplicate signs that can have custom values but also have a version that cannot have one, like the maxspeed/speedzone signs.

The solution is to add a static sign svg filename to the state object and build it before we modify the sign object based on the custom value.

`DE?signs=DE:274.1[50];265[3.5],1042-35[Su,PH+222];Eingeschränkter+Winterdienst`